requirements.txt
SQLAlchemy
fastapi
pydantic
opencv-python
numpy
pytesseract
pillow
openai
onnxruntime
requests
python-dotenv
uvicorn
python-multipart
opencv-python
opencv-python-headless
numpy
psycopg
starlette